requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
supabase
pdfminer.six
python-docx
